+++ README.md
@@ -1,6 +1,6 @@
-Fork of AcousticChat for IronForge Server.
+Fork of org.acoustic.chat.AcousticChat for IronForge Server.
 
-Changes from AcousticChat:
+Changes from org.acoustic.chat.AcousticChat:
 - Maven Support - DONE
 - VentureChat Support - DONE
 
@@ -8,4 +8,4 @@ Credits to
 
 https://github.com/cs-au-dk/dk.brics.automaton - Dependency <br>
 https://github.com/mifmif/Generex - Dependency<br>
-https://github.com/Cabbache/AcousticChat - Original 
+https://github.com/Cabbache/org.acoustic.chat.AcousticChat - Original
