--- src/main/java/external/dk/brics/automaton/package.html
@@ -1,17 +1,0 @@
-<html>
-<body>
-Finite-state automaton for regular expressions.
-<p>
-This package contains a full DFA/NFA implementation with Unicode
-alphabet and support for all standard (and a number of non-standard)
-regular expression operations.
-<p>
-The most commonly used functionality is located in the classes
-<tt><a href="Automaton.html">Automaton</a></tt> and
-<tt><a href="RegExp.html">RegExp</a></tt>.
-<p>
-For more information, go to the package home page at 
-<tt><a href="http://www.brics.dk/automaton/" 
-target="_top">http://www.brics.dk/automaton/</a></tt>.
-</body>
-</html>
